requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
black==21.12b0
certifi==2021.10.8
click==8.0.3
Django==4.0
django-durationwidget==1.0.5
django-enumfields==2.1.1
django-filter==2.4.0
django-jazzmin==2.4.8
django-json-widget==1.0.1
djangorestframework==3.13.1
djangorestframework-filters==1.0.0.dev2
factory-boy==2.12.0
Faker==11.1.0
future==0.18.2
importlib-metadata==4.10.0
Markdown==3.3.6
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.1
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tomli==1.2.3
typing_extensions==4.0.1
zipp==3.7.0
